refactor: refatorando method de login user
