--- src/main/java/br/com/guedes/apartments/models/dto/AuthResponse.java
@@ -1,4 +1,0 @@
-package br.com.guedes.apartments.models.dto;
-
-public record AuthResponse(String token) {
-}
