--- src/main/java/br/com/guedes/apartments/models/dto/AuthenticationDTO.java
@@ -1,3 +1,0 @@
-package br.com.guedes.apartments.models.dto;
-
-public record AuthenticationDTO(String cpf, String password) {}
