--- src/main/java/br/com/guedes/apartments/models/dto/RegisterDTO.java
@@ -1,5 +1,0 @@
-package br.com.guedes.apartments.models.dto;
-
-import br.com.guedes.apartments.models.enums.Role;
-
-public record RegisterDTO(String cpf, String password, Role role) {}
