+++ src/main/java/br/com/guedes/apartments/models/dto/requests/RequestLoginDTO.java
@@ -1,0 +1,3 @@
+package br.com.guedes.apartments.models.dto.requests;
+
+public record RequestLoginDTO(String cpf, String password) {}
