+++ src/main/java/br/com/guedes/apartments/models/dto/requests/RequestRegisterDTO.java
@@ -1,0 +1,5 @@
+package br.com.guedes.apartments.models.dto.requests;
+
+import br.com.guedes.apartments.models.enums.Role;
+
+public record RequestRegisterDTO(String cpf, String password, Role role) {}
