+++ src/main/java/br/com/guedes/apartments/models/dto/responses/DefaultResponseDTO.java
@@ -1,3 +1,3 @@
-package br.com.guedes.apartments.models.dto;
+package br.com.guedes.apartments.models.dto.responses;
 
 public record DefaultResponseDTO(int code, String message) {}
