+++ src/main/java/br/com/guedes/apartments/models/dto/responses/LoginTokenResponseDTO.java
@@ -1,0 +1,4 @@
+package br.com.guedes.apartments.models.dto.responses;
+
+public record LoginTokenResponseDTO(String token, String description) {
+}
