+++ src/main/java/br/com/guedes/apartments/models/enums/Message.java
@@ -2,7 +2,8 @@ package br.com.guedes.apartments.models.enums;
 
 public enum Message {
 
-    CREATED_USER(201, "SUCCESSFULLY REGISTERED USER"),
+    CREATED_USER(201, "Successfully registered user"),
+    LOGIN_AUTHENTICATED(201, "Token generated successfully, you can browse the app"),
     DUPLICATED_KEY(409, "CPF already registered");
     private final int code;
     private final String description;
