+++ src/main/java/br/com/guedes/apartments/services/UserService.java
@@ -2,7 +2,6 @@ package br.com.guedes.apartments.services;
 
 import br.com.guedes.apartments.dao.UserDAO;
 import br.com.guedes.apartments.models.UserRequest;
-import br.com.guedes.apartments.models.dto.RegisterDTO;
 import br.com.guedes.apartments.models.dto.UserResponse;
 import br.com.guedes.apartments.models.dto.UserSecurityDetails;
 import org.springframework.beans.factory.annotation.Autowired;
